remove underscore in class names
