+++ warp10/src/main/java/io/warp10/script/WarpScriptLib.java
@@ -58,10 +58,10 @@ import io.warp10.script.aggregator.TrueCourse;
 import io.warp10.script.aggregator.VDist;
 import io.warp10.script.aggregator.VSpeed;
 import io.warp10.script.aggregator.Variance;
-import io.warp10.script.aggregator.modifier.NULLS_FAIL;
-import io.warp10.script.aggregator.modifier.NULLS_KEEP;
-import io.warp10.script.aggregator.modifier.NULLS_REMOVE;
-import io.warp10.script.aggregator.modifier.NULLS_SKIP;
+import io.warp10.script.aggregator.modifier.NULLSFAIL;
+import io.warp10.script.aggregator.modifier.NULLSKEEP;
+import io.warp10.script.aggregator.modifier.NULLSREMOVE;
+import io.warp10.script.aggregator.modifier.NULLSSKIP;
 import io.warp10.script.binary.ADD;
 import io.warp10.script.binary.BitwiseAND;
 import io.warp10.script.binary.BitwiseOR;
@@ -2035,10 +2035,10 @@ public class WarpScriptLib {
     addNamedWarpScriptFunction(new MACROFILLER(MACROFILLER));
     addNamedWarpScriptFunction(new STRICTMAPPER(STRICTMAPPER));
     addNamedWarpScriptFunction(new STRICTREDUCER(STRICTREDUCER));
-    addNamedWarpScriptFunction(new NULLS_FAIL(NULLS_FAIL));
-    addNamedWarpScriptFunction(new NULLS_KEEP(NULLS_KEEP));
-    addNamedWarpScriptFunction(new NULLS_REMOVE(NULLS_REMOVE));
-    addNamedWarpScriptFunction(new NULLS_SKIP(NULLS_SKIP));
+    addNamedWarpScriptFunction(new NULLSFAIL(NULLS_FAIL));
+    addNamedWarpScriptFunction(new NULLSKEEP(NULLS_KEEP));
+    addNamedWarpScriptFunction(new NULLSREMOVE(NULLS_REMOVE));
+    addNamedWarpScriptFunction(new NULLSSKIP(NULLS_SKIP));
     addNamedWarpScriptFunction(new PARSESELECTOR(PARSESELECTOR));
     addNamedWarpScriptFunction(new TOSELECTOR(TOSELECTOR));
     addNamedWarpScriptFunction(new PARSE(PARSE));
