+++ warp10/src/main/java/io/warp10/script/aggregator/modifier/NULLSKEEP.java
@@ -18,7 +18,7 @@ package io.warp10.script.aggregator.modifier;
 
 import io.warp10.continuum.gts.Aggregate;
 import io.warp10.script.NamedWarpScriptFunction;
-import io.warp10.script.WarpScriptAggregatorFailIfAnyNull;
+import io.warp10.script.WarpScriptAggregatorKeepNulls;
 import io.warp10.script.WarpScriptException;
 import io.warp10.script.WarpScriptLib;
 import io.warp10.script.WarpScriptReducer;
@@ -26,8 +26,8 @@ import io.warp10.script.WarpScriptStack;
 import io.warp10.script.WarpScriptStackFunction;
 import io.warp10.script.functions.SNAPSHOT;
 
-public class NULLS_FAIL extends NamedWarpScriptFunction implements WarpScriptStackFunction {
-  public NULLS_FAIL(String name) {
+public class NULLSKEEP extends NamedWarpScriptFunction implements WarpScriptStackFunction {
+  public NULLSKEEP(String name) {
     super(name);
   }
 
@@ -39,28 +39,27 @@ public class NULLS_FAIL extends NamedWarpScriptFunction implements WarpScriptSta
       throw new WarpScriptException(getName() + " expects a reducer");
     }
 
-    if (!(o instanceof WarpScriptAggregatorFailIfAnyNull)) {
+    if (!(o instanceof WarpScriptAggregatorKeepNulls)) {
       throw new WarpScriptException(getName() + " cannot be be applied to this AGGREGATOR");
     }
 
-    stack.push(new ModifiedAggregator(getName(), (WarpScriptAggregatorFailIfAnyNull) o));
+    stack.push(new ModifiedAggregator(getName(), (WarpScriptAggregatorKeepNulls) o));
 
     return stack;
   }
 
-  private static final class ModifiedAggregator extends NamedWarpScriptFunction implements WarpScriptReducer, WarpScriptAggregatorFailIfAnyNull, SNAPSHOT.Snapshotable {
+  private static final class ModifiedAggregator extends NamedWarpScriptFunction implements WarpScriptReducer, WarpScriptAggregatorKeepNulls, SNAPSHOT.Snapshotable {
 
-    private final WarpScriptAggregatorFailIfAnyNull aggregator;
+    private final WarpScriptAggregatorKeepNulls aggregator;
 
-    public ModifiedAggregator(String name, WarpScriptAggregatorFailIfAnyNull aggregator) {
+    public ModifiedAggregator(String name, WarpScriptAggregatorKeepNulls aggregator) {
       super(name);
       this.aggregator = aggregator;
     }
 
     @Override
     public Object apply(Aggregate aggregate) throws WarpScriptException {
-      WarpScriptAggregatorFailIfAnyNull.failIfAnyNull(aggregate);
-      return aggregator.apply(aggregate);
+      return aggregator.apply(WarpScriptAggregatorKeepNulls.keepNulls(aggregate));
     }
 
     @Override
