+++ warp10/src/main/java/io/warp10/script/aggregator/modifier/NULLSSKIP.java
@@ -17,8 +17,9 @@
 package io.warp10.script.aggregator.modifier;
 
 import io.warp10.continuum.gts.Aggregate;
+import io.warp10.continuum.gts.GeoTimeSerie;
 import io.warp10.script.NamedWarpScriptFunction;
-import io.warp10.script.WarpScriptAggregatorKeepNulls;
+import io.warp10.script.WarpScriptAggregatorSkipIfAnyNull;
 import io.warp10.script.WarpScriptException;
 import io.warp10.script.WarpScriptLib;
 import io.warp10.script.WarpScriptReducer;
@@ -26,8 +27,8 @@ import io.warp10.script.WarpScriptStack;
 import io.warp10.script.WarpScriptStackFunction;
 import io.warp10.script.functions.SNAPSHOT;
 
-public class NULLS_KEEP extends NamedWarpScriptFunction implements WarpScriptStackFunction {
-  public NULLS_KEEP(String name) {
+public class NULLSSKIP extends NamedWarpScriptFunction implements WarpScriptStackFunction {
+  public NULLSSKIP(String name) {
     super(name);
   }
 
@@ -39,27 +40,30 @@ public class NULLS_KEEP extends NamedWarpScriptFunction implements WarpScriptSta
       throw new WarpScriptException(getName() + " expects a reducer");
     }
 
-    if (!(o instanceof WarpScriptAggregatorKeepNulls)) {
+    if (!(o instanceof WarpScriptAggregatorSkipIfAnyNull)) {
       throw new WarpScriptException(getName() + " cannot be be applied to this AGGREGATOR");
     }
 
-    stack.push(new ModifiedAggregator(getName(), (WarpScriptAggregatorKeepNulls) o));
+    stack.push(new ModifiedAggregator(getName(), (WarpScriptAggregatorSkipIfAnyNull) o));
 
     return stack;
   }
 
-  private static final class ModifiedAggregator extends NamedWarpScriptFunction implements WarpScriptReducer, WarpScriptAggregatorKeepNulls, SNAPSHOT.Snapshotable {
+  private static final class ModifiedAggregator extends NamedWarpScriptFunction implements WarpScriptReducer, WarpScriptAggregatorSkipIfAnyNull, SNAPSHOT.Snapshotable {
 
-    private final WarpScriptAggregatorKeepNulls aggregator;
+    private final WarpScriptAggregatorSkipIfAnyNull aggregator;
 
-    public ModifiedAggregator(String name, WarpScriptAggregatorKeepNulls aggregator) {
+    public ModifiedAggregator(String name, WarpScriptAggregatorSkipIfAnyNull aggregator) {
       super(name);
       this.aggregator = aggregator;
     }
 
     @Override
     public Object apply(Aggregate aggregate) throws WarpScriptException {
-      return aggregator.apply(WarpScriptAggregatorKeepNulls.keepNulls(aggregate));
+      if (WarpScriptAggregatorSkipIfAnyNull.containsAnyNull(aggregate)) {
+        return new Object[] { Long.MAX_VALUE, GeoTimeSerie.NO_LOCATION, GeoTimeSerie.NO_ELEVATION, null };
+      }
+      return aggregator.apply(aggregate);
     }
 
     @Override
